requirements.txt
# Data fetching
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0

# Jupyter for research
jupyter>=1.0.0
notebook>=7.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# News APIs
finnhub-python>=2.4.0
requests>=2.31.0
feedparser>=6.0.0

# Discord notifications
discord-webhook>=1.3.0

# Schwab API (when ready)
# schwab-py>=1.0.0

# Scheduling
schedule>=1.2.0

# Environment management
python-dotenv>=1.0.0
